chore: 폴더 생성 및 함수 분리
