+++ src/app/mypage/_types/type.ts
@@ -9,18 +9,19 @@ export type TabContents = {
   reviews_written: Reviews[];
   reviews_unwritten: Reviews[];
 };
-export type Items = {
-  thumbnail: string;
-  title: string;
-};
 
 export type Orders = {
-  order_id: number;
   amount: number;
   created_at: string;
   item_id: number;
-  order_status: string;
   items: Items;
+  order_id: number;
+  order_status: string;
+};
+
+export type Items = {
+  thumbnail: string;
+  title: string;
 };
 
 export type Reviews = {
