+++ src/hooks/mutate/useUpdateStatus.ts
@@ -1,6 +1,6 @@
-import { OrderStatus } from '@/components/shop-manage/order/temp';
-import supabase from '@/services/supabase';
+import { OrderStatus } from '@/types/orderType';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { updateOrderStatus } from '@/api/orderedItemsApi';
 
 /** 주문 상품 상태를 변경하는 쿼리 훅 */
 export const useUpdateOrderStatus = () => {
@@ -17,13 +17,4 @@ export const useUpdateOrderStatus = () => {
       console.error(error.message);
     },
   });
-};
-
-// 아이템의 상태 바꾸기
-export const updateOrderStatus = async (orderId: number, status: OrderStatus) => {
-  const { error } = await supabase
-    .from('ordered_items')
-    .update({ order_status: status })
-    .eq('order_id', orderId);
-  if (error) throw new Error('상품 상태를 바꾸는 도중 오류가 발생했습니다.');
 };
