+++ src/utils/translateOrderStatus.ts
@@ -1,14 +1,4 @@
-import { Item } from '@/types/cartType';
-
-// 주문 아이템 페이지에서 사용할 아이템 타입
-export type OrderedItem = {
-  amount: number;
-  created_at: string;
-  item_id: number;
-  item: Item;
-  order_id: number;
-  order_status: OrderStatus;
-};
+import { OrderStatus } from '@/types/orderType';
 
 // 주문 상태를 한국어로 변환
 export const ORDER_STATUS = (status: OrderStatus): string => {
@@ -25,5 +15,3 @@ export const ORDER_STATUS = (status: OrderStatus): string => {
       return '취소';
   }
 };
-
-export type OrderStatus = 'pending' | 'paid' | 'shipped' | 'delivered' | 'cancelled';
